Changed from using Agent as keys to AgentID as keys to the orderMap. However, will probably need to change it back for easier implementation of the checkMakeTrade() method.
Partially implemented checkMakeTrade() (when an agent makes an order
that matches sell/buy price). Still have to modify code on modifyOrder
so that it can trigger a trade. Need code to notify agent when trade has
occurred and log the action.

Implemented methods to get top ten sell/buy orders based on price and
time creation.